// src/generated/types/CourseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CourseDto = {
  /**
   * @type string
   */
  name: string
  /**
   * @type number
   */
  creditHours: number
  /**
   * @type string
   */
  subjectCode: string
  /**
   * @type number
   */
  courseNumber: number
  /**
   * @type boolean
   */
  hasLab: boolean
  /**
   * @type string
   */
  labDuration: string
  /**
   * @type number
   */
  attendanceMarks: number
  /**
   * @type string
   */
  id: string
  /**
   * @type string, date-time
   */
  created_at: Date
  /**
   * @type string, date-time
   */
  updated_at: Date
}